Los sistemas binarios (de momento) retienen su posición.
* Corregido un error de sobreescritura que ocurría al recargar los datos
  guardados, pues stellar_neighbourhood los volvía a generar. A partir
  de esta revisión, los datos posicionales sólo se generan 1 vez, y se
  retienen siempre que se recargue el mismo archivo de datos.
* Corregido un  crash en el StarPanel.
* La cuenta de los sistemas binarios por crear, en el panel de sistemas
  estelares, ahora cuenta correctmente los sistemas cargados desde el
  savefile.
* Eliminada una función inutil en StellarNeighbourhood y agregada otra
  para el preprocesamiento de datos.
* Acomodado un poco el texto de la Assertion en binary.system_type()

+++ engine/frontend/widgets/panels/star_panel.py
@@ -113,7 +113,8 @@ class StarPanel(BasePanel):
     def add_button(self, star):
         button = StarButton(self.current, star, str(star), self.curr_x, self.curr_y)
         self.properties.add(button, layer=2)
-        self.proto_stars.delete_objects(self.proto_stars.current.object_data)
+        if len(self.proto_stars):
+            self.proto_stars.delete_objects(self.proto_stars.current.object_data)
         Systems.add_star(star)
         if star not in self.stars:
             self.stars.append(star)
